+++ kleidicv_thread/include/kleidicv_thread/kleidicv_thread.h
@@ -398,6 +398,18 @@ kleidicv_error_t kleidicv_thread_remap_s16point5_u8(
     kleidicv_border_type_t border_type, kleidicv_border_values_t border_values,
     kleidicv_thread_multithreading);
 
+/// Internal - not part of the public API and its direct use is not supported.
+///
+/// Multithreaded implementation of kleidicv_warp_perspective_u8 - see the
+/// documentation of that function for more details.
+kleidicv_error_t kleidicv_thread_warp_perspective_u8(
+    const uint8_t *src, size_t src_stride, size_t src_width, size_t src_height,
+    uint8_t *dst, size_t dst_stride, size_t dst_width, size_t dst_height,
+    const float transformation[9], size_t channels,
+    kleidicv_interpolation_type_t interpolation,
+    kleidicv_border_type_t border_type, kleidicv_border_values_t border_values,
+    kleidicv_thread_multithreading);
+
 #ifdef __cplusplus
 }  // extern "C"
 #endif  // __cplusplus