+++ scripts/ci-opencv.sh
@@ -102,6 +102,8 @@ IMGPROC_TEST_PATTERNS=(
   '*Imgproc_Dilate*'
   '*Imgproc_Erode*'
   '*Imgproc_PyramidDown*'
+  '*Imgproc_Remap*'
+  '*Imgproc_Warp*'
 )
 IMGPROC_TEST_PATTERNS_STR="$(join_strings_with_colon "${IMGPROC_TEST_PATTERNS[*]}")"
 ../../../conformity/opencv_kleidicv/bin/opencv_test_imgproc \